+++ packages/oauth/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.4.0
+
+- [Breaking] Require `lucia-auth` 0.4.3
+- Log request errors on dev mode
+
 ## 0.3.2
 
 - [Fix] Fix runtime errors